Deleted school submission file. Added file with school extras removed and jupyter cell syntax removed.
